Integrate file into `index.ts`.
